requirements.txt
tensorflow==2.13.1
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.13.1
matplotlib
numpy
